Injected security layer into code.

+++ src/server/model/query_strings.js
@@ -12,7 +12,10 @@ export const QUERY_STRINGS = {
       FROM mmc.movie_genre AS g
       JOIN mmc.movie AS m
         ON g.movie_id = m.id
-     WHERE g.genre = $1::mmc.genre`,
+      JOIN mmc.user_movie AS u
+        ON u.movie_id = m.id
+     WHERE g.genre = $1::mmc.genre
+       AND u.user_id = $2`,
 
   // Talent Queries
   SELECT_ALL_TALENT: 'SELECT * FROM mmc.talent',
@@ -22,7 +25,10 @@ export const QUERY_STRINGS = {
       FROM mmc.movie_talent AS t
       JOIN mmc.movie AS m
         ON t.movie_id = m.id
-     WHERE t.talent_id = $1`,
+      JOIN mmc.user_movie AS u
+        ON u.movie_id = m.id
+     WHERE t.talent_id = $1
+       AND u.user_id = $2`,
   INSERT_TALENT: `
     INSERT INTO mmc.talent (first_name, last_name)
     VALUES($1, $2)
@@ -30,18 +36,34 @@ export const QUERY_STRINGS = {
   `,
 
   // Movie Queries
-  SELECT_ALL_MOVIES: 'SELECT id, title, poster FROM mmc.movie',
-  SELECT_ONE_MOVIE: 'SELECT * FROM mmc.movie WHERE id = $1',
-  SELECT_MOVIE_GENRE: 'SELECT genre FROM mmc.movie_genre WHERE movie_id = $1',
+  SELECT_ALL_MOVIES: `
+    SELECT id, title, poster 
+      FROM mmc.movie
+     WHERE user_id = $1`,
+  SELECT_ONE_MOVIE: `
+    SELECT * 
+      FROM mmc.movie 
+     WHERE id = $1
+       AND user_id = $2)`,
+  SELECT_MOVIE_GENRE: `
+    SELECT genre 
+      FROM mmc.movie_genre AS g
+      JOIN mmc.movie AS m
+        ON m.id = g.movie_id
+     WHERE movie_id = $1
+       AND user_id = $2`,
   SELECT_MOVIE_TALENT: `
     SELECT t.*
       FROM mmc.talent AS t
       JOIN mmc.movie_talent AS m
         ON m.talent_id = t.id
-     WHERE m.movie_id = $1`,
+      JOIN mmc.movie AS u
+        ON u.id = m.movie_id
+     WHERE m.movie_id = $1
+       AND u.user_id = $2`,
   INSERT_MOVIE: `
-    INSERT INTO mmc.movie (title, release_year, rating, poster)
-    VALUES ($1, $2, $3, $4)
+    INSERT INTO mmc.movie (title, release_year, rating, poster, user_id)
+    VALUES ($1, $2, $3, $4, $5)
     RETURNING *`,
   UPDATE_MOVIE: `
     UPDATE mmc.movie
@@ -50,21 +72,38 @@ export const QUERY_STRINGS = {
           ,rating = $4
           ,poster = $5
      WHERE id = $1
+       AND user_id = $6
     RETURNING *
   `,
-  DELETE_MOVIE: 'DELETE FROM mmc.movie WHERE id = $1',
+  DELETE_MOVIE: 'DELETE FROM mmc.movie WHERE id = $1 AND user_id = $2',
   ADD_TALENT_TO_MOVIE: `
     INSERT INTO mmc.movie_talent (movie_id, talent_id)
-    VALUES ($1, $2)
+    SELECT $1, $2
+      FROM mmc.movie
+     WHERE id = $1
+       AND user_id = $3
     RETURNING *`,
   DELETE_TALENT_FROM_MOVIE: `
     DELETE FROM mmc.movie_talent
      WHERE movie_id = $1
-       AND talent_id = $2`,
+       AND talent_id = $2
+       AND EXISTS (
+         SELECT *
+           FROM mmc.movie
+          WHERE id = movie_id
+            AND user_id = $3`,
   ADD_GENRE_TO_MOVIE: `
     INSERT INTO mmc.movie_genre (movie_id, genre)
-    VALUES ($1, $2)
+    SELECT $1, $2
+      FROM mmc.movie
+     WHERE id = $1
+       AND user_id = $3
     RETURNING *`,
-  DELETE_GENRE_FROM_MOVIE:
-    'DELETE FROM mmc.movie_genre WHERE movie_id = $1 AND genre = $2',
+  DELETE_GENRE_FROM_MOVIE: `DELETE FROM mmc.movie_genre 
+      WHERE movie_id = $1 AND genre = $2
+        AND EXISTS (
+          SELECT *
+            FROM mmc.movie
+           WHERE id = movie_id
+             AND user_id = $3`,
 };
